* David Snopek's MIDI recording on AMD64 patch and Gentoo build patch
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6246 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sequencer/rosegardensequencer.h
@@ -103,7 +103,7 @@ public:
 	       const Rosegarden::RealTime &audioRead,
 	       const Rosegarden::RealTime &audioWrite,
 	       long smallFileSize,
-               int recordMode);
+               long recordMode);
 
     // looping
     void setLoop(const Rosegarden::RealTime &loopStart,
@@ -137,7 +137,7 @@ public:
 		       long audioWriteSec,
 		       long audioWriteNsec,
 		       long smallFileSize,
-                       int recordMode);
+                       long recordMode);
 
     
     // Jump to a pointer in the playback (uses longs instead
